# Optimise getentropy